Fix the horde manager overlay with non-1 GUI scale
With GUI scale above 1 the overlay will be drawn off-screen. GUI scale below 1 (which requires additional mods or manually editing setting files) would likely result in the overlay being drawn away from the edge, moved towards the left side of the screen.

The fix is to use `UI.screenWidth` for screen width, as it's the value RimWorld calculates based on the current GUI scale.